refactor(capsule): extract topic matching into dedicated module
Move has_valid_segments and topic_matches from dispatcher.rs into
topic.rs. Pure extraction — no logic changes. Reduces dispatcher.rs
from 1072 to 936 lines.

Also addresses review: explicitly handle all InterceptResult variants
in the single-interceptor dispatch path instead of Ok(_).
